+++ db/updates/update_12.4-12.5.sql
@@ -118,3 +118,11 @@ CREATE TABLE `sdsSettings` (
   `updated_at` TIMESTAMP on update CURRENT_TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP , 
   PRIMARY KEY (`id`)
 ) ENGINE = InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
+
+ALTER TABLE `formulasMetaData` ADD INDEX(`owner_id`);
+ALTER TABLE `formulas` ADD INDEX(`owner_id`);
+ALTER TABLE `formulas` ADD INDEX(`fid`);
+ALTER TABLE `formulas` ADD INDEX(`ingredient`);
+ALTER TABLE `ingredients` ADD INDEX(`owner_id`);
+ALTER TABLE `ingredients` ADD INDEX(`name`);
+