+++ pages/listIngredients.php
@@ -80,6 +80,7 @@ $cIngredients = mysqli_num_rows(mysqli_query($conn, "SELECT id FROM ingredients 
    </thead>
 </table>
 
+<li><a class="dropdown-item" id="json_export" href="#" data-bs-toggle="modal" data-bs-target="#export_options_modal"><i class="fa-solid fa-file-code mx-2"></i>Export to JSON</a></li>
 
 <!-- Modal -->
 <div class="modal fade" id="export_options_modal" tabindex="-1" aria-labelledby="exportOptionsLabel" aria-hidden="true">
@@ -227,7 +228,7 @@ $(document).ready(function() {
 			data: function(d) {
 				d.pvSearch =  '<?=$_GET['search']?>'
 				d.provider = $('#pv_search_btn').attr('data-provider')
-				d.adv = '<?=$_GET['adv']?:0?>'
+				d.advanced = '<?=$_GET['advanced']?:0?>'
 				d.profile = '<?=$_GET['profile']?:null?>'
 				d.name = '<?=$_GET['name']?:null?>'
 				d.cas = '<?=$_GET['cas']?:null?>'
@@ -464,7 +465,7 @@ $(document).ready(function() {
 					'<ul class="dropdown-menu dropdown-menu-right">';
 			data += '<li><a href="/pages/mgmIngredient.php?id=' + row.id + '" class="dropdown-item popup-link"><i class="fas fa-edit mx-2"></i>Manage</a></li>';
 			
-			data += '<li><a class="dropdown-item" href="/pages/export.php?format=json&kind=single-ingredient&id=' + row.id + '" rel="tip" title="Export '+ row.name +' as JSON" ><i class="fas fa-download mx-2"></i>Export to JSON</a></li>';
+			data += '<li><a class="dropdown-item" href="/pages/export.php?format=json&kind=single-ingredient&id=' + row.id + '" rel="tip" title="Export '+ row.name +' as JSON" ><i class="fas fa-download mx-2"></i>Export as JSON</a></li>';
 	
 			data += '<div class="dropdown-divider"></div>';