+++ pages/views/users/users.php
@@ -249,7 +249,7 @@ $(document).ready(function() {
 		'<button type="button" class="btn btn-floating hidden-arrow" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="bi bi-three-dots-vertical"></i></button>' +
 			'<ul class="dropdown-menu">';	
         data += '<li><a class="dropdown-item" href="#" data-bs-target="#editUser" data-bs-toggle="editUser" id="editUser" rel="tip" title="Edit ' + row.full_name + '" data-id=' + row.id + ' data-name="' + row.full_name +'"><i class="bi bi-pencil-square mx-2"></i>Edit</a></li>';
-        if (row.id !== <?php echo $userID; ?>) {
+        if (row.id !== "<?php echo $userID; ?>") {
             data += '<li><a class="dropdown-item" href="#" id="impersonateUser" rel="tip" title="Impersonate ' + row.full_name + '" data-id=' + row.id + ' data-name="' + row.full_name + '"><i class="bi bi-person-bounding-box mx-2"></i>Impersonate</a></li>';
         }
 		data += '<li><a class="dropdown-item text-danger" href="#" id="deleteUser" rel="tip" title="Delete ' + row.full_name + '" data-id=' + row.id + ' data-name="' + row.full_name + '"><i class="bi bi-trash mx-2"></i>Delete</a></li>';